Backend/requirements.txt
fastapi==0.95.1
uvicorn==0.20.0
pydantic==1.10.4
pydantic[email]==1.10.4
requests==2.28.1
sqlalchemy==1.4.48
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==4.6.0
